refactor(site): move landing page files to src/site and swap routes
