+++ src/site/StaticDiffApp.test.tsx
