+++ src/site/types/staticDiff.ts
