+++ src/site/utils/staticApiBridge.ts
